Huge amount of optimsation work on new ignition system

+++ speeduino/crankMaths.h
@@ -11,12 +11,12 @@
 #define fastDegreesToUS(targetDegrees) (((targetDegrees) * (unsigned long)timePerDegreex16) >> 4)
 #define fastTimeToAngle(time) (((unsigned long)time * degreesPeruSx2048) / 2048) //Divide by 2048 will be converted at compile time to bitshift
 
-#define ignitionLimits(angle) (angle >= CRANK_ANGLE_MAX_IGN ? (angle - CRANK_ANGLE_MAX_IGN) : (angle < 0 ? (angle + CRANK_ANGLE_MAX_IGN) : angle))
+#define ignitionLimits(angle) ( ((int16_t)angle >= CRANK_ANGLE_MAX_IGN) ? (angle - CRANK_ANGLE_MAX_IGN) : ( (angle < 0) ? (angle + CRANK_ANGLE_MAX_IGN) : angle))
 
 unsigned long angleToTime(int16_t, byte);
 uint16_t timeToAngle(unsigned long, byte);
 
-volatile int timePerDegree;
+volatile uint16_t timePerDegree;
 volatile uint16_t timePerDegreex16;
 volatile uint16_t degreesPeruSx2048;
